# cmd/backend/docs/backend_swagger.yaml
definitions:
  trx-project_internal_model.Permission:
    properties:
      action:
        description: 操作：read, write, delete
        type: string
      code:
        description: 权限编码：user:read, user:write
        type: string
      created_at:
        type: string
      description:
        description: 权限描述
        type: string
      id:
        type: integer
      name:
        description: 权限名称：查看用户、编辑用户
        type: string
      resource:
        description: 资源：user, role, permission, statistics
        type: string
      status:
        description: 状态：1-启用 0-禁用
        type: integer
      updated_at:
        type: string
    type: object
  trx-project_internal_model.Role:
    properties:
      created_at:
        type: string
      description:
        description: 角色描述
        type: string
      display_name:
        description: 显示名称：超级管理员、管理员、编辑、查看者
        type: string
      id:
        type: integer
      name:
        description: 角色名称：superadmin, admin, editor, viewer
        type: string
      permissions:
        description: 关联
        items:
          $ref: '#/definitions/trx-project_internal_model.Permission'
        type: array
      status:
        description: 状态：1-启用 0-禁用
        type: integer
      updated_at:
        type: string
    type: object
...
  trx-project_pkg_response.Response:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 数据
      message:
        description: 消息
        type: string
      trace_id:
        description: 追踪ID（可选）
        type: string
    type: object
info:
  contact: {}
paths:
  /admin/rbac/permissions:
    get:
      consumes:
      - application/json
      description: 获取所有权限列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取权限列表
          schema:
            allOf:
            - $ref: '#/definitions/trx-project_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/trx-project_internal_model.Permission'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取权限列表
      tags:
      - RBAC管理
  /admin/rbac/roles:
    get:
      consumes:
      - application/json
      description: 获取所有角色列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取角色列表
          schema:
            allOf:
            - $ref: '#/definitions/trx-project_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/trx-project_internal_model.Role'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取角色列表
      tags:
      - RBAC管理
    post:
      consumes:
      - application/json
      description: 创建新的角色
      parameters:
      - description: 角色信息
        in: body
        name: request
        required: true
        schema:
          properties:
            description:
              type: string
            display_name:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/trx-project_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/trx-project_internal_model.Role'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "409":
          description: 角色名已存在
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 创建角色
      tags:
      - RBAC管理
  /admin/rbac/roles/{id}:
    get:
      consumes:
      - application/json
      description: 获取指定角色的详细信息，包括权限列表
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取角色详情
          schema:
            allOf:
            - $ref: '#/definitions/trx-project_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/trx-project_internal_model.Role'
              type: object
        "400":
          description: 无效的角色ID
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取角色详情
      tags:
      - RBAC管理
  /admin/rbac/roles/{id}/permissions:
    post:
      consumes:
      - application/json
      description: 为指定角色分配权限列表
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限ID列表
        in: body
        name: request
        required: true
        schema:
          properties:
            permission_ids:
              items:
                type: integer
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 分配成功
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 为角色分配权限
      tags:
      - RBAC管理
...
  /admin/users/{id}/permissions:
    get:
      consumes:
      - application/json
      description: 获取指定用户的所有权限（通过角色继承）
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取用户权限
          schema:
            allOf:
            - $ref: '#/definitions/trx-project_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/trx-project_internal_model.Permission'
                  type: array
              type: object
        "400":
          description: 无效的用户ID
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户权限
      tags:
      - RBAC管理
...
  /admin/users/{id}/role:
    post:
      consumes:
      - application/json
      description: 为指定用户分配角色
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色ID
        in: body
        name: request
        required: true
        schema:
          properties:
            role_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 分配成功
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 为用户分配角色
      tags:
      - RBAC管理
  /admin/users/{id}/roles:
    get:
      consumes:
      - application/json
      description: 获取指定用户的角色列表
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取用户角色
          schema:
            allOf:
            - $ref: '#/definitions/trx-project_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/trx-project_internal_model.Role'
                  type: array
              type: object
        "400":
          description: 无效的用户ID
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户角色
      tags:
      - RBAC管理
...
swagger: "2.0"